Rename mqtt-monitor.py to mqtt_monitor.py for valid module name
