+++ .github/workflows/build.yml
@@ -12,7 +12,7 @@ jobs:
       - name: Create .deb structure
         run: |
           mkdir -p mqtt-monitor_deb/{DEBIAN,usr/bin,etc/mqtt-monitor,lib/systemd/system,usr/share/mqtt-monitor}
-          cp src/mqtt-monitor.py mqtt-monitor_deb/usr/bin/
+          cp src/mqtt_monitor.py mqtt-monitor_deb/usr/bin/
           cp src/mqtt-monitor.conf mqtt-monitor_deb/usr/share/mqtt-monitor/
           cp src/merge_config.py mqtt-monitor_deb/usr/share/mqtt-monitor/
           cp debian/mqtt-monitor.service mqtt-monitor_deb/lib/systemd/system/
@@ -23,7 +23,7 @@ jobs:
           cp debian/postinst mqtt-monitor_deb/DEBIAN/
           cp debian/prerm mqtt-monitor_deb/DEBIAN/
           cp debian/postrm mqtt-monitor_deb/DEBIAN/
-          chmod 755 mqtt-monitor_deb/usr/bin/mqtt-monitor.py
+          chmod 755 mqtt-monitor_deb/usr/bin/mqtt_monitor.py
           chmod 644 mqtt-monitor_deb/usr/share/mqtt-monitor/mqtt-monitor.conf
           chmod 644 mqtt-monitor_deb/usr/share/mqtt-monitor/merge_config.py
           chmod 644 mqtt-monitor_deb/lib/systemd/system/mqtt-monitor.service
