+++ .github/workflows/release.yml
@@ -17,7 +17,7 @@ jobs:
           pip install pytest
           pip install -r requirements.txt
       - name: Run tests
-        run: pytest tests/test_mqtt-monitor.py
+        run: pytest tests/test_mqtt_monitor.py
       - name: Get version from tag
         id: get_version
         run: echo "VERSION=${GITHUB_REF#refs/tags/v}" >> $GITHUB_ENV
@@ -26,7 +26,7 @@ jobs:
       - name: Create .deb structure
         run: |
           mkdir -p mqtt-monitor_${{ env.VERSION }}/{DEBIAN,usr/bin,etc/mqtt-monitor,lib/systemd/system,usr/share/mqtt-monitor}
-          cp src/mqtt-monitor.py mqtt-monitor_${{ env.VERSION }}/usr/bin/
+          cp src/mqtt_monitor.py mqtt-monitor_${{ env.VERSION }}/usr/bin/
           cp src/mqtt-monitor.conf mqtt-monitor_${{ env.VERSION }}/usr/share/mqtt-monitor/
           cp src/merge_config.py mqtt-monitor_${{ env.VERSION }}/usr/share/mqtt-monitor/
           cp debian/mqtt-monitor.service mqtt-monitor_${{ env.VERSION }}/lib/systemd/system/
@@ -37,7 +37,7 @@ jobs:
           cp debian/postinst mqtt-monitor_${{ env.VERSION }}/DEBIAN/
           cp debian/prerm mqtt-monitor_${{ env.VERSION }}/DEBIAN/
           cp debian/postrm mqtt-monitor_${{ env.VERSION }}/DEBIAN/
-          chmod 755 mqtt-monitor_${{ env.VERSION }}/usr/bin/mqtt-monitor.py
+          chmod 755 mqtt-monitor_${{ env.VERSION }}/usr/bin/mqtt_monitor.py
           chmod 644 mqtt-monitor_${{ env.VERSION }}/usr/share/mqtt-monitor/mqtt-monitor.conf
           chmod 644 mqtt-monitor_${{ env.VERSION }}/usr/share/mqtt-monitor/merge_config.py
           chmod 644 mqtt-monitor_${{ env.VERSION }}/lib/systemd/system/mqtt-monitor.service
