+++ .github/workflows/test.yaml
@@ -17,4 +17,4 @@ jobs:
           pip install pytest
           pip install -r requirements.txt
       - name: Run tests
-        run: pytest tests/test_mqtt-monitor.py
+        run: pytest tests/test_mqtt_monitor.py
