+++ tests/Dockerfile
@@ -13,5 +13,3 @@ RUN pip3 install -r requirements.txt pytest
 
 COPY src/ /app/src/
 COPY tests/ /app/tests/
-
-ENV PYTHONPATH=/app/src
